+++ src/main/java/com/pax/market/api/sdk/java/api/constant/Constants.java
@@ -70,7 +70,7 @@ public class Constants {
     public static final int DEFAULT_PAGE_SIZE = 10;
     
     public static final String THIRD_PARTY_API_SDK_LANGUAGE = "Java";
-    public static final String THIRD_PARTY_API_SDK_VERSION = "8.6.0";
+    public static final String THIRD_PARTY_API_SDK_VERSION = "8.7.0";
     public static final String REQ_HEADER_SDK_LANG = "SDK-Language";
     public static final String REQ_HEADER_SDK_VERSION = "SDK-Version";