+++ src/main/java/com/pax/market/api/sdk/java/api/merchantVariable/dto/MerchantVariableUpdateRequest.java
@@ -1,0 +1,24 @@
+/*
+ * *******************************************************************************
+ * COPYRIGHT
+ *               PAX TECHNOLOGY, Inc. PROPRIETARY INFORMATION
+ *   This software is supplied under the terms of a license agreement or
+ *   nondisclosure agreement with PAX  Technology, Inc. and may not be copied
+ *   or disclosed except in accordance with the terms in that agreement.
+ *
+ *      Copyright (C) 2022 PAX Technology, Inc. All rights reserved.
+ * *******************************************************************************
+ */
+
+package com.pax.market.api.sdk.java.api.merchantVariable.dto;
+
+import com.pax.market.api.sdk.java.api.variable.dto.ParameterVariable;
+
+
+/**
+ * @author shifan
+ * @date 2022/10/19
+ */
+public class MerchantVariableUpdateRequest extends ParameterVariable {
+    private static final long serialVersionUID = -5178222104284377409L;
+}