+++ src/main/java/com/pax/market/api/sdk/java/api/terminalVariable/TerminalVariableApi.java
@@ -23,12 +23,12 @@ import com.pax.market.api.sdk.java.api.terminalVariable.dto.*;
 import com.pax.market.api.sdk.java.api.util.CryptoUtils;
 import com.pax.market.api.sdk.java.api.util.EnhancedJsonUtils;
 import com.pax.market.api.sdk.java.api.util.StringUtils;
+import com.pax.market.api.sdk.java.api.variable.dto.ParameterVariable;
+import com.pax.market.api.sdk.java.api.variable.dto.ParameterVariableDTO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.security.GeneralSecurityException;
 import java.util.List;
 import java.util.Locale;
 
@@ -55,7 +55,7 @@ public class TerminalVariableApi  extends BaseThirdPartySysApi {
         super(baseUrl, apiKey, apiSecret, locale);
     }
 
-    public Result<ParameterVariableDTO> getTerminalVariable(int pageNo, int pageSize , SearchOrderBy orderBy,String tid , String serialNo , String packageName, String key, VariableSource source){
+    public Result<ParameterVariableDTO> getTerminalVariable(int pageNo, int pageSize , SearchOrderBy orderBy, String tid , String serialNo , String packageName, String key, VariableSource source){
         ThirdPartySysApiClient client = new ThirdPartySysApiClient(getBaseUrl(), getApiKey(), getApiSecret());
         PageRequestDTO page = new PageRequestDTO();
         page.setPageNo(pageNo);