+++ docs/CHANGELOG.md
@@ -1,11 +1,15 @@
 # Changelog
 
+## 8.7.0
+### New Feature
+* Add merchant variable APIs
+* terminal API support copy a exist terminal
+
+
 ## 8.6.0
 ### Improvement
 * The fields such as effectiveTime and expiredTime are added respectively for terminal and group push related tasks.
 * Search terminals API is updated, add resellerName and merchantName parameters.
-
-
 
 
 ## 8.5.1